+++ R/bbk.R
@@ -69,7 +69,7 @@ bbk_data = function(
   updated_after = NULL
 ) {
   assert_string(flow, min.chars = 5L, max.chars = 8L)
-  assert_character(key, min.chars = 1L, null.ok = TRUE)
+  assert_character(key, min.chars = 1L, min.len = 1L, any.missing = FALSE, null.ok = TRUE)
   assert_period(start_period)
   assert_period(end_period)
   first_n = assert_count(first_n, null.ok = TRUE, positive = TRUE, coerce = TRUE)